Make it possible to force the running of autogen.
With the change to run autogen on make all if there
was no makefile present, it became impossible to
force the running of autogen when that is needed. Fix
that by adding a bootstrap-autogen target that checks
the existing of Makefiles and reverting the autogen
target to always run autogen.