Allow for direct entry of group names when creating delegations.
This requires a bit of trickery. I use the onblur() javascript function
to note when the field is left and store whatever was entered there.

Then when the page is submitted if a dn doesn't exist for that field
but they did enter something, do a lookup to see if there is a group by
that name.