In the UI we don't want to display Edit links unless someone can actually edit things. We use the 'editors' group for this. This group itself grants no permission other than displaying certain things in the UI.
In order to be in the editors group a user must be a member of a group that
is the source group in a delegation. The memberof plugin will do all the
hard work to be sure that a user's memberof contains cn=editors if they
are in a delegated group.

432874