Under some conditions rl may not have been initialized so the config may error out with:
UnboundLocalError: "local variable 'rl' referenced before assignment"

This is caught and ignored but the result is that the records in DNS may
not be used at all. Initializing rl to zero fixes this.

I also convert the server list into a set to make each entry unique (and
back to a list because that is what we are supposed to return)

433506